requirements.txt
fbchat-asyncio
prompt_toolkit
